pyproject.toml
[tool.poetry]
name = "dhos-questions-api"
version = "0.0.1"
description = "Questions API to manage questionnaires and feedback"
authors = ["Andrew Gill <andrew.gill@sensynehealth.com>"]
keywords = ["Questions", "Answers", "DHOS"]
repository = "https://github.com/draysontechnologies/dhos-questions-api"

[[tool.poetry.source]]
name = "sensynehealth"
url = "https://pypi.fury.io/sensynehealth/"

[tool.poetry.dependencies]
python = "^3.9"
she-logging = "1.*"
flask-batteries-included = {version = "3.*", extras = ["apispec", "pgsql"]}

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coloredlogs = "*"
coverage = "*"
isort = "*"
mock = "*"
mypy = "*"
pytest = "*"
pytest-dhos = {version = "*", extras=["fbi"]}
pytest-flask = "*"
pytest-freezegun = "*"
pytest-mock = "*"
requests_mock = "*"
sadisplay = "*"
safety = "*"
tox = "*"
types-mock = "*"
types-PyYAML ="*"
types-waitress = "*"

[tool.mypy]
namespace_packages=true
ignore_missing_imports=false
disallow_untyped_defs=true

[[tool.mypy.overrides]]
module = [
    "waitress",
    "pytest",
    "flask.testing",
    "apispec.*",
    "connexion",
    "apispec_webframeworks.*",
    "pytest_mock",
    "jose",
    "flask",
    "sadisplay",
    "pytest_dhos.*",
    "sqlalchemy.*",
    "flask_sqlalchemy.*"
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
known_third_party = ["alembic", "apispec", "apispec_webframeworks", "behave", "click", "clients", "connexion", "environs", "faker", "flask", "flask_batteries_included", "flask_sqlalchemy", "helpers", "jose", "marshmallow", "mock", "pytest", "pytest_dhos", "pytest_mock", "reporting", "reportportal_behave", "requests", "sadisplay", "she_logging", "sqlalchemy", "waitress", "yaml"]

[tool.black]
line-length = 88
target-version = ["py39"]
